refactor: make InstructionCard and PredictionCard common components
