+++ src/components/ui/features/common/PredictionCard.jsx
